+++ packages/labs/gen-wrapper-vue/src/index.ts
@@ -15,7 +15,6 @@ import {wrapperModuleTemplateSFC} from './lib/wrapper-module-template-sfc.js';
 import {FileTree} from '@lit-labs/gen-utils/lib/file-utils.js';
 import {tsconfigNodeTemplate} from './lib/tsconfig.node-template.js';
 import {viteConfigTemplate} from './lib/vite.config-template.js';
-import {renameTemplate} from './lib/rename-template.js';
 
 export const getCommand = () => {
   return {
@@ -52,7 +51,6 @@ export const generateVueWrapper = async (pkg: Package): Promise<FileTree> => {
         'tsconfig.json': tsconfigTemplate(),
         'tsconfig.node.json': tsconfigNodeTemplate(),
         'vite.config.ts': viteConfigTemplate(pkg.packageJson, sfcFiles),
-        'scripts/rename.js': renameTemplate(),
         ...sfcFiles,
       },
     };