+++ packages/labs/testing/src/test/good-element_test.ts
@@ -28,7 +28,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
 
     test('renders with default values', async () => {
       const el = await fixture(html`<good-element></good-element>`, {
-        base: import.meta.url,
         modules: ['./good-element.js'],
       });
       assert.shadowDom.equal(
@@ -45,7 +44,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
       const el = await fixture(
         html`<good-element name="Test"></good-element>`,
         {
-          base: import.meta.url,
           modules: ['./good-element.js'],
         }
       );
@@ -63,7 +61,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
       const el = await fixture<GoodElement>(
         html`<good-element></good-element>`,
         {
-          base: import.meta.url,
           modules: ['./good-element.js'],
         }
       );
@@ -76,7 +73,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
         const el = await fixture<GoodElement>(
           html`<good-element></good-element>`,
           {
-            base: import.meta.url,
             modules: ['./good-element.js'],
           }
         );
@@ -98,7 +94,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
   suite(`nested good-element rendered with ${fixture.name}`, () => {
     test('renders with default values', async () => {
       const el = await fixture(html`<div><good-element></good-element></div>`, {
-        base: import.meta.url,
         modules: ['./good-element.js'],
       });
       const goodEl = el.querySelector('good-element');
@@ -117,7 +112,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
       const el = await fixture(
         html`<div><good-element name="Test"></good-element></div>`,
         {
-          base: import.meta.url,
           modules: ['./good-element.js'],
         }
       );
@@ -133,7 +127,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
 
     test('styling applied', async () => {
       const el = await fixture(html`<div><good-element></good-element></div>`, {
-        base: import.meta.url,
         modules: ['./good-element.js'],
       });
       assert.equal(
@@ -148,7 +141,6 @@ for (const fixture of [csrFixture, ssrNonHydratedFixture, ssrHydratedFixture]) {
         const el = await fixture(
           html`<div><good-element></good-element></div>`,
           {
-            base: import.meta.url,
             modules: ['./good-element.js'],
           }
         );