+++ packages/lit-html/src/directive.ts
@@ -33,7 +33,8 @@ export type DirectiveParameters<C extends Directive> = Parameters<C['render']>;
 export interface DirectiveResult<C extends DirectiveClass = DirectiveClass> {
   /**
    * This property needs to remain unminified.
-   * @internal */
+   * @internal
+   */
   ['_$litDirective$']: C;
   /** @internal */
   values: DirectiveParameters<InstanceType<C>>;
@@ -123,7 +124,7 @@ export abstract class Directive implements Disconnectable {
   _$initialize(
     part: Part,
     parent: Disconnectable,
-    attributeIndex: number | undefined,
+    attributeIndex: number | undefined
   ) {
     this.__part = part;
     this._$parent = parent;