+++ packages/lit-html/src/directives/class-map.ts
@@ -37,7 +37,7 @@ class ClassMapDirective extends Directive {
     ) {
       throw new Error(
         '`classMap()` can only be used in the `class` attribute ' +
-          'and must be the only part in the attribute.',
+          'and must be the only part in the attribute.'
       );
     }
   }
@@ -62,7 +62,7 @@ class ClassMapDirective extends Directive {
           part.strings
             .join(' ')
             .split(/\s/)
-            .filter((s) => s !== ''),
+            .filter((s) => s !== '')
         );
       }
       for (const name in classInfo) {
@@ -111,7 +111,7 @@ class ClassMapDirective extends Directive {
  * This must be used in the `class` attribute and must be the only part used in
  * the attribute. It takes each property in the `classInfo` argument and adds
  * the property name to the element's `classList` if the property value is
- * truthy; if the property value is falsey, the property name is removed from
+ * truthy; if the property value is falsy, the property name is removed from
  * the element's `class`.
  *
  * For example `{foo: bar}` applies the class `foo` if the value of `bar` is