+++ packages/lit-html/src/static.ts
@@ -7,7 +7,12 @@
 // Any new exports need to be added to the export statement in
 // `packages/lit/src/index.all.ts`.
 
-import {html as coreHtml, svg as coreSvg, TemplateResult} from './lit-html.js';
+import {
+  html as coreHtml,
+  svg as coreSvg,
+  mathml as coreMathml,
+  TemplateResult,
+} from './lit-html.js';
 
 export interface StaticValue {
   /** The value to interpolate as-is into the template. */
@@ -15,7 +20,7 @@ export interface StaticValue {
 
   /**
    * A value that can't be decoded from ordinary JSON, make it harder for
-   * a attacker-controlled data that goes through JSON.parse to produce a valid
+   * an attacker-controlled data that goes through JSON.parse to produce a valid
    * StaticValue.
    */
   r: typeof brand;
@@ -71,7 +76,7 @@ const textFromStatic = (value: StaticValue) => {
   } else {
     throw new Error(
       `Value passed to 'literal' function must be a 'literal' result: ${value}. Use 'unsafeStatic' to pass non-literal values, but
-            take care to ensure page security.`,
+            take care to ensure page security.`
     );
   }
 };
@@ -96,7 +101,7 @@ export const literal = (
 ): StaticValue => ({
   ['_$litStatic$']: values.reduce(
     (acc, v, idx) => acc + textFromStatic(v as StaticValue) + strings[idx + 1],
-    strings[0],
+    strings[0]
   ) as string,
   r: brand,
 });
@@ -107,7 +112,7 @@ const stringsCache = new Map<string, TemplateStringsArray>();
  * Wraps a lit-html template tag (`html` or `svg`) to add static value support.
  */
 export const withStatic =
-  (coreTag: typeof coreHtml | typeof coreSvg) =>
+  (coreTag: typeof coreHtml | typeof coreSvg | typeof coreMathml) =>
   (strings: TemplateStringsArray, ...values: unknown[]): TemplateResult => {
     const l = values.length;
     let staticValue: string | undefined;
@@ -155,7 +160,7 @@ export const withStatic =
         (staticStrings as any).raw = staticStrings;
         stringsCache.set(
           key,
-          (strings = staticStrings as unknown as TemplateStringsArray),
+          (strings = staticStrings as unknown as TemplateStringsArray)
         );
       }
       values = dynamicValues;
@@ -178,3 +183,11 @@ export const html = withStatic(coreHtml);
  * Includes static value support from `lit-html/static.js`.
  */
 export const svg = withStatic(coreSvg);
+
+/**
+ * Interprets a template literal as MathML fragment that can efficiently render
+ * to and update a container.
+ *
+ * Includes static value support from `lit-html/static.js`.
+ */
+export const mathml = withStatic(coreMathml);