+++ packages/localize/examples/transform-ts/rollup.config.mjs
@@ -1,6 +1,6 @@
 import typescript from '@rollup/plugin-typescript';
 import resolve from '@rollup/plugin-node-resolve';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
 import {summary} from 'rollup-plugin-summary';
 import {localeTransformers} from '@lit/localize-tools/lib/rollup.js';