+++ packages/labs/observers/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lit-labs/observers
 
+## 2.0.3
+
+### Patch Changes
+
+- [#4591](https://github.com/lit/lit/pull/4591) [`045b6f15`](https://github.com/lit/lit/commit/045b6f159815edb9e690bc1f6829d467f42aa520) Thanks [@a11delavar](https://github.com/a11delavar)! - Do not initialize observers to prevent failing in SSR environment.
+
+- Updated dependencies [[`feccc1ba`](https://github.com/lit/lit/commit/feccc1ba8e82b36d07a0e2576381bf2819926b98)]:
+  - lit-html@3.2.0
+
 ## 2.0.2
 
 ### Patch Changes